requirements.txt
discordwebhook
python-dotenv